feat(vfs) Continue the rewrite of the in-memory filesystem.
