+++ lib/wasi/src/lib.rs
@@ -12,6 +12,27 @@
 //! [WASI plugin example](https://github.com/wasmerio/wasmer/blob/master/examples/plugin.rs)
 //! for an example of how to extend WASI using the WASI FS API.
 
+#[cfg(all(not(feature = "sys"), not(feature = "js")))]
+compile_error!("At least the `sys` or the `js` feature must be enabled. Please, pick one.");
+
+#[cfg(all(feature = "sys", feature = "js"))]
+compile_error!(
+    "Cannot have both `sys` and `js` features enabled at the same time. Please, pick one."
+);
+
+#[cfg(all(feature = "sys", target_arch = "wasm32"))]
+compile_error!("The `sys` feature must be enabled only for non-`wasm32` target.");
+
+#[cfg(all(feature = "js", not(target_arch = "wasm32")))]
+compile_error!(
+    "The `js` feature must be enabled only for the `wasm32` target (either `wasm32-unknown-unknown` or `wasm32-wasi`)."
+);
+
+#[cfg(all(feature = "host-fs", feature = "mem-fs"))]
+compile_error!(
+    "Cannot have both `host-fs` and `mem-fs` features enabled at the same time. Please, pick one."
+);
+
 #[macro_use]
 mod macros;
 mod ptr;
